pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cascade"
version = "0.1.0"
description = "Local multi-agent mandate runner"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
cascade = "cascade.cli:main"

[tool.setuptools]
packages = ["cascade"]

[tool.pytest.ini_options]
testpaths = ["tests"]